record git AuthorDate and time zone in bzr revision properties
and use them when re-importing the revision into bzr instead of
the usual CommitDate and time zone.

This solves following problem:
if you have two independent git repositories synced with one bzr
branch and from one of those, you push a git revision that is
the result of a cherry-pick merge or other operation that gives
it a different commit and author date, then pull the bzr revision
into the other git repository, it will arrive there altered because
both dates enter the sha1 hash.