more roundtrip security for author, committer and message
Turns out the way committer/author names and emails are processed
is also not roundtrip safe, fixing that by attaching the exact lines
from git as commit properties.

The commit message is also mangled sometimes: bzr does not like \r
characters in anything, and sometimes git exports the commit message
without trailing newline. So in case we won't be able to reconstruct
the full message, we attach it as base64 encoded in another property.